refactor(core): rename the converter file to schema-draft.functions
The file no longer holds one function that converts to draft 6. It holds
draft detection as well, and the split adds a third, so the name has
stopped describing the contents.

Follows the directory convention: a plural .functions.ts for a module of
several, matching layout.functions.ts and jsonpointer.functions.ts, where
merge-schemas.function.ts and clone-deep.function.ts stay singular.

The public API is untouched. shared/index.ts re-exports the same symbol
name and only its path changes, confirmed by typechecking an import of
convertSchemaToDraft6 from the built package.
