+++ projects/ajsf-core/src/lib/shared/index.ts
@@ -28,7 +28,7 @@ export {
   resolveSchemaReferences, getSubSchema, combineAllOf, fixRequiredArrayProperties
 } from './json-schema.functions';
 
-export { convertSchemaToDraft6 } from './convert-schema-to-draft6.function';
+export { convertSchemaToDraft6 } from './schema-draft.functions';
 
 export { mergeSchemas } from './merge-schemas.function';
